Updated readme and new camera presets
