+++ camera_presets/Hero_7_4K_60_16by9_wide_.json
@@ -1,0 +1,40 @@
+{
+    "name": "Camera name",
+    "note": "",
+    "calibrator_version": "0.1.0-alpha",
+    "date": "2020-10-05",
+    "calib_dimension": {
+        "w": 4000,
+        "h": 3000
+    },
+    "num_images": 5,
+    "use_opencv_fisheye": true,
+    "fisheye_params": {
+        "RMS_error": 4.026367568950084,
+        "camera_matrix": [
+            [
+                1768.7130910058195,
+                0.0,
+                2000.0
+            ],
+            [
+                0.0,
+                1769.1791034052221,
+                1500.0
+            ],
+            [
+                0.0,
+                0.0,
+                1.0
+            ]
+        ],
+        "distortion_coeffs": [
+            0.06874491798328626,
+            -0.1539301797100358,
+            0.5679016436800388,
+            -0.6420654738068317
+        ]
+    },
+    "use_opencv_standard": false,
+    "calib_params": {}
+}
