+++ test scripts/crisp_testing.py
